adds player direction and inferno

+++ backend/parser/models/grenade_explosion.go
@@ -3,5 +3,5 @@ package parsermodels
 // GrenadeExplosion schema
 type GrenadeExplosion struct {
 	Frame     int
-	GrenadeID int
+	GrenadeID int64
 }
